time in reward service now a double

+++ src/main/java/com/quantumslots/physhacks/service/RewardService.java
@@ -3,7 +3,7 @@ package com.quantumslots.physhacks.service;
 import com.quantumslots.physhacks.model.Potentials;
 
 public class RewardService {
-    public int getReward(Potentials potential, float selector1, float selector2, float time, int bet){
+    public int getReward(Potentials potential, float selector1, float selector2, double time, int bet){
         switch (potential){
             case InfiniteSquareWell:
                 return -1;
